### What this PR does

Previously varlogadm and varlogmr defined the flag for replication factor, respectively, which made strings for the flag different.
This change adds a new standard flag, `--replication-factor`. It also sets the type of replication factor to an integer.

+++ tests/it/testenv.go
@@ -1200,7 +1200,7 @@ func (clus *VarlogCluster) initVMS(t *testing.T) {
 	opts := append(clus.VMSOpts,
 		admin.WithListenAddress(listenAddress),
 		admin.WithClusterID(clus.clusterID),
-		admin.WithReplicationFactor(uint(clus.nrRep)),
+		admin.WithReplicationFactor(clus.nrRep),
 		admin.WithLogger(clus.logger.Named("admin")),
 		admin.WithMetadataRepositoryManager(mrMgr),
 		admin.WithStorageNodeManager(snMgr),
